Validation of profile photos, and resizing them to thumbnails
This verifies the initial bytes of the file match the magic bytes of pngs, jpegs, gifs or bmps, and constructs a tightly-constrained call to imagemagick's `convert` to downsize the image to a jpeg of at most 200x300 pixels and at most 40kb in file-size.  If this succeeds, the initial uploaded file is deleted and the user's profile is updated to the thumbnail.  This helps speed up roster pages (which show students' thumbnails on popover) and the course facebook page (which shows everyone in the course) that have to download a lot of images...