Fix NPE in grader_allocations/index and /edit, in the following scenario:
* Create a team submission for an assignment with a manual grader
* Allocate a grader to that submission, and have the grader enter some comments but not complete the grading
* Split the submission into individual subs
* Try to view or edit the grader allocations for the assignment now

The problem is that manual_grader would guess that a reasonable grader for the new subs exists, but because grading isn't finished, this should not be true.  The fix is two-fold: harden the guessing so that it only checks for completed grades (this grade.score.nil? check is now the same in both ManualGrader and index.html, so these stay in sync), and fix the sub-splitting process to migrate any existing grader allocations to the new submissions.