Better fix for broken uploads.  When creating a new assignment, with new graders that need uploaded files, we run into the same circularity as with assignments having starter files.
This turns out to be quite hard to fix, because we don't create the graders explicitly, but rather do so through assigning bulk attributes.  This in turn causes us to change how uploads save their data, to make it part of their transactional .save method.  This had some knock-on effects in assignments, graders, and submissions; and affected the test factories too.  Hopefully this is now a more stable design and will work reliably.